# openapi parsing performance improvement with protobuffer