--- a/api/Makefile
+++ b/api/Makefile
@@ -5,7 +5,7 @@ include ../Makefile-modules.mk
 
 test:
 	go test -v -timeout 45m -cover ./... -ldflags "-X sigs.k8s.io/kustomize/api/provenance.version=v444.333.222"
-	cd api/krusty; OPENAPI_TEST=true go test -run TestCustomOpenAPIFieldFromComponentWithOverlays
+	cd krusty/openapitests; OPENAPI_TEST=true go test -v -timeout 45m -p 1 -cover ./...
 
 build:
 	go build -ldflags "-X sigs.k8s.io/kustomize/api/provenance.version=v444.333.222" ./...