--- a/cmd/config/internal/commands/cmdset_test.go
+++ b/cmd/config/internal/commands/cmdset_test.go
@@ -502,8 +502,8 @@ spec:
   list: # {"$ref":"#/definitions/io.k8s.cli.setters.list"}
   - 0
  `,
-			errMsg: `list in body must be of type integer: "string"
-list in body must be of type integer: "boolean"
+			errMsg: `list[1] in body must be of type integer: "string"
+list[2] in body must be of type integer: "boolean"
 list in body should have at most 2 items`,
 		},
 
@@ -751,8 +751,8 @@ spec:
   list: # {"$ref":"#/definitions/io.k8s.cli.setters.list"}
   - 0
  `,
-			errMsg: `list in body must be of type integer: "string"
-list in body must be of type integer: "boolean"
+			errMsg: `list[1] in body must be of type integer: "string"
+list[2] in body must be of type integer: "boolean"
 list in body should have at most 2 items`,
 		},
 		{