--- a/kyaml/openapi/kubernetesapi/v1218pb/doc.go
+++ b/kyaml/openapi/kubernetesapi/v1218pb/doc.go
@@ -1,6 +1,0 @@
-// Copyright 2022 The Kubernetes Authors.
-// SPDX-License-Identifier: Apache-2.0
-
-package v1218pb
-
-const OpenAPIProtobuf = "v1218pb"