… api/go.sum
… api/krusty/openapicustomschema_test.go
… api/krusty/openapiversion_test.go
… cmd/config/go.sum
… cmd/pluginator/go.sum
… cmd/pluginator/internal/krmfunction/funcwrapper/statik.go
… kustomize/go.sum
… kyaml/go.sum
… kyaml/openapi/kubernetesapi/v1212/swagger.go
… kyaml/openapi/kubernetesapi/v1212/swagger.pb
… kyaml/openapi/kubernetesapi/v1218pb/swagger.go
… kyaml/openapi/kubernetesapi/v1218pb/swagger.pb
… plugin/builtin/annotationstransformer/go.sum
… plugin/builtin/configmapgenerator/go.sum
… plugin/builtin/hashtransformer/go.sum
… plugin/builtin/helmchartinflationgenerator/go.sum
… plugin/builtin/iampolicygenerator/go.sum
… plugin/builtin/imagetagtransformer/go.sum
… plugin/builtin/labeltransformer/go.sum
… plugin/builtin/legacyordertransformer/go.sum
… plugin/builtin/namespacetransformer/go.sum
… plugin/builtin/patchjson6902transformer/go.sum
… plugin/builtin/patchstrategicmergetransformer/go.sum
… plugin/builtin/patchtransformer/go.sum
… plugin/builtin/prefixtransformer/go.sum
… plugin/builtin/replacementtransformer/go.sum
… plugin/builtin/replicacounttransformer/go.sum
… plugin/builtin/secretgenerator/go.sum
… plugin/builtin/suffixtransformer/go.sum
… plugin/builtin/valueaddtransformer/go.sum