Fixes https://github.com/kubernetes-sigs/kustomize/issues/2987 
Fixes https://github.com/kubernetes-sigs/kustomize/issues/4100 
Fixes https://github.com/kubernetes-sigs/kustomize/issues/3670

Related: https://github.com/kubernetes-sigs/kustomize/issues/4569

The performance improvement changes the format we use to parse our builtin openapi from JSON to proto. See https://github.com/kubernetes/kube-openapi/pull/283, it reduces the parsing time from `~628ms` to `~39ms`.

You can see the performance improvement when running the example provided [on this issue](https://github.com/kubernetes-sigs/kustomize/issues/2987#issuecomment-956070256):

Previously:

```
$ time kustomize build
apiVersion: bitnami.com/v1alpha1
kind: SealedSecret
metadata:
  name: example

real    0m0.670s
user    0m0.688s
sys     0m0.037s
```

With this branch:

```
$ time kustomize build
apiVersion: bitnami.com/v1alpha1
kind: SealedSecret
metadata:
  name: example

real    0m0.109s
user    0m0.114s
sys     0m0.029s
```

You can also see about a half second improvement in performance when running certain tests, e.g.

On master:
```
$ time go test -run TestPatchesInOneFile
PASS
ok      sigs.k8s.io/kustomize/api/krusty        0.902s

real    0m2.407s
user    0m2.540s
sys     0m1.120s
```

On this branch:
```
$ time go test -run TestPatchesInOneFile
PASS
ok      sigs.k8s.io/kustomize/api/krusty        0.343s

real    0m1.845s
user    0m1.990s
sys     0m1.144s
```

There is a separate PR to update the openapi scripts here: https://github.com/kubernetes-sigs/kustomize/pull/4582