@natasha41575 2022-04-05
> Do the openapi update scripts still work? They still have a bunch of swagger.json references. There are also a handful of comments referring to swagger.json that I think need to be changed.

That's a good point. I'll update the scripts as well and audit kyaml for any other references to swagger.json.